Searched text is done in BankPlugin only & revert changes made in BankTagsPlugin.
Remove setBankTitle from onScriptCallbackEvent

After BANKMAIN_BUILD runs, grab all visible items and set the bank title on it.
