+++ runelite-client/src/main/java/net/runelite/client/plugins/banktags/BankTagsConfig.java
@@ -54,39 +54,6 @@ public interface BankTagsConfig extends Config
 	}
 
 	@ConfigItem(
-		keyName = "showGE",
-		name = "Show Grand Exchange price",
-		description = "Show grand exchange price total (GE)",
-		position = 3
-	)
-	default boolean showGE()
-	{
-		return true;
-	}
-
-	@ConfigItem(
-		keyName = "showHA",
-		name = "Show high alchemy price",
-		description = "Show high alchemy price total (HA)",
-		position = 4
-	)
-	default boolean showHA()
-	{
-		return false;
-	}
-
-	@ConfigItem(
-		keyName = "showExact",
-		name = "Show exact bank value",
-		description = "Show exact bank value",
-		position = 5
-	)
-	default boolean showExact()
-	{
-		return false;
-	}
-
-	@ConfigItem(
 		keyName = "position",
 		name = "",
 		description = "",
